Corrected and excluded sample project packages from coverage
