--- samples/spring-boot-web/src/main/java/org/apache/shiro/examples/HelloController.java
@@ -1,61 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.shiro.examples;
-
-import org.apache.shiro.SecurityUtils;
-import org.apache.shiro.subject.PrincipalCollection;
-import org.apache.shiro.subject.Subject;
-import org.apache.shiro.util.CollectionUtils;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import javax.servlet.http.HttpServletRequest;
-import java.util.Collection;
-import java.util.Map;
-
-@Controller
-public class HelloController {
-
-    @SuppressWarnings("Duplicates")
-    @RequestMapping("/")
-    public String home(HttpServletRequest request, Model model) {
-
-        String name = "World";
-
-        Subject subject = SecurityUtils.getSubject();
-
-        PrincipalCollection principalCollection = subject.getPrincipals();
-
-        if (principalCollection != null && !principalCollection.isEmpty()) {
-            Collection<Map> principalMaps = subject.getPrincipals().byType(Map.class);
-            if (CollectionUtils.isEmpty(principalMaps)) {
-                name = subject.getPrincipal().toString();
-            }
-            else {
-                name = (String) principalMaps.iterator().next().get("username");
-            }
-        }
-
-        model.addAttribute("name", name);
-
-        return "hello";
-    }
-
-}
