+++ samples/spring-boot-web/src/main/java/org/apache/shiro/samples/AccountInfoController.java
@@ -16,37 +16,37 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.shiro.examples;
+package org.apache.shiro.samples;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.web.ErrorAttributes;
-import org.springframework.boot.autoconfigure.web.ErrorController;
+import org.apache.shiro.SecurityUtils;
+import org.apache.shiro.authz.annotation.RequiresRoles;
+import org.apache.shiro.subject.PrincipalCollection;
+import org.apache.shiro.subject.Subject;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.context.request.ServletRequestAttributes;
 
-import javax.servlet.http.HttpServletRequest;
-import java.util.Map;
 
-/**
- */
 @Controller
-public class RestrictedErrorController implements ErrorController {
-    private static final String ERROR_PATH = "/error";
+public class AccountInfoController {
 
-    @Autowired
-    private ErrorAttributes errorAttributes;
+    @RequiresRoles("admin")
+    @RequestMapping("/account-info")
+    public String home(Model model) {
 
-    @Override
-    public String getErrorPath() {
-        return ERROR_PATH;
+        String name = "World";
+
+        Subject subject = SecurityUtils.getSubject();
+
+        PrincipalCollection principalCollection = subject.getPrincipals();
+
+        if (principalCollection != null && !principalCollection.isEmpty()) {
+            name = principalCollection.getPrimaryPrincipal().toString();
+        }
+
+        model.addAttribute("name", name);
+
+        return "account-info";
     }
 
-    @RequestMapping(ERROR_PATH)
-    String error(HttpServletRequest request, Model model) {
-        Map<String, Object> errorMap = errorAttributes.getErrorAttributes(new ServletRequestAttributes(request), false);
-        model.addAttribute("errors", errorMap);
-        return "error";
-    }
 }
