+++ samples/spring-boot-web/src/main/java/org/apache/shiro/samples/LoginController.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.shiro.examples;
+package org.apache.shiro.samples;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
